@@ -1,11 +1,11 @@
 name: syntax_highlighter
 description: syntax highlighter for show your dart code in the application.
-version: 0.1.0
+version: 0.1.1
 author: Tushar Parmar <viztushar@gmail.com>
 homepage: https://github.com/viztushar/syntax_highlighter
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
